requirements.txt
discord.py
python-dotenv
discord_webhook
SQLAlchemy